+++ src/main/java/nl/mitwnn/ch17/ctrl_z/receptopzak/controller/IngredientController.java
@@ -1,13 +1,15 @@
 package nl.mitwnn.ch17.ctrl_z.receptopzak.controller;
 
 import nl.mitwnn.ch17.ctrl_z.receptopzak.model.Ingredient;
+import nl.mitwnn.ch17.ctrl_z.receptopzak.model.Recipe;
 import nl.mitwnn.ch17.ctrl_z.receptopzak.repositories.IngredientRepository;
+import nl.mitwnn.ch17.ctrl_z.receptopzak.repositories.RecipeRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-
+import java.util.Optional;
 
 /**
  * @author Pelle Meuzelaar
@@ -19,9 +21,11 @@ import org.springframework.web.bind.annotation.*;
 public class IngredientController {
 
     private final IngredientRepository ingredientRepository;
+    private final RecipeRepository recipeRepository;
 
-    public IngredientController(IngredientRepository ingredientRepository) {
+    public IngredientController(IngredientRepository ingredientRepository, RecipeRepository recipeRepository) {
         this.ingredientRepository = ingredientRepository;
+        this.recipeRepository = recipeRepository;
     }
 
     // Show ingredients
@@ -37,17 +41,29 @@ public class IngredientController {
     @PostMapping("/save")
     public String saveOrUpdateIngredient(@ModelAttribute("formIngredient") Ingredient ingredient, BindingResult result) {
         if (result.hasErrors()) {
-            return "redirect:/ingredients/all";
+            return "redirect:/ingredient/all";
         }
 
+        ingredient.setIngredientKcal();
         ingredientRepository.save(ingredient);
-        return "redirect:/ingredients/all";
+        return "redirect:/ingredient/all";
     }
 
     // Delete ingredients
     @GetMapping("/delete/{ingredientId}")
     public String deleteIngredient(@PathVariable("ingredientId") Long ingredientId) {
-        ingredientRepository.deleteById(ingredientId);
+        Optional<Ingredient> optionalIngredient = ingredientRepository.findById(ingredientId);
+
+        if (optionalIngredient.isPresent()) {
+            Ingredient ingredient = optionalIngredient.get();
+
+            for (Recipe recipe : recipeRepository.findAll()) {
+                recipe.getIngredients().remove(ingredient);
+                recipeRepository.save(recipe);
+            }
+
+            ingredientRepository.deleteById(ingredientId);
+        }
         return "redirect:/ingredient/all";
     }
 }
